update of cordis harvesting code
